[SYCL][Graph] Unify tests which uses common header kernel functions
- Unify tests which use common header kernel functions that are mostly identical between APIs
- Test code moved to Inputs folder and actual lit tests define an API and include the input file
- Added missing tests for some APIs which have a counterpart in the other and are not API specific
- New common functions which are API agnostic for adding diamond dependency nodes
- New API agnostic function for adding single nodes
- API selected by defining GRAPH_E2E_<EXPLICIT/RECORD_REPLAY> before including test input
