+++ sycl/test-e2e/Graph/Explicit/temp_buffer_reinterpret.cpp
@@ -8,4 +8,4 @@
 
 #define GRAPH_E2E_EXPLICIT
 
-#include "../Inputs/basic_usm.cpp"
+#include "../Inputs/temp_buffer_reinterpret.cpp"
