+++ sycl/test-e2e/Graph/RecordReplay/multiple_exec_graphs.cpp
@@ -1,73 +1,11 @@
 // REQUIRES: level_zero, gpu
 // RUN: %{build} -o %t.out
 // RUN: %{run} %t.out
+// Extra run to check for leaks in Level Zero using ZE_DEBUG
 // RUN: %if ext_oneapi_level_zero %{env ZE_DEBUG=4 %{run} %t.out 2>&1 | FileCheck %s %}
 //
 // CHECK-NOT: LEAK
 
-// This test attempts creating multiple executable graphs from one modifiable
-// graph.
-// The second run is to check that there are no leaks reported with the embedded
-// ZE_DEBUG=4 testing capability.
+#define GRAPH_E2E_RECORD_REPLAY
 
-#include "../graph_common.hpp"
-
-int main() {
-  queue Queue;
-
-  using T = int;
-
-  std::vector<T> DataA(Size), DataB(Size), DataC(Size);
-
-  std::iota(DataA.begin(), DataA.end(), 1);
-  std::iota(DataB.begin(), DataB.end(), 10);
-  std::iota(DataC.begin(), DataC.end(), 1000);
-
-  std::vector<T> ReferenceA(DataA), ReferenceB(DataB), ReferenceC(DataC);
-  calculate_reference_data(Iterations, Size, ReferenceA, ReferenceB,
-                           ReferenceC);
-
-  exp_ext::command_graph Graph{Queue.get_context(), Queue.get_device()};
-
-  T *PtrA = malloc_device<T>(Size, Queue);
-  T *PtrB = malloc_device<T>(Size, Queue);
-  T *PtrC = malloc_device<T>(Size, Queue);
-
-  Queue.copy(DataA.data(), PtrA, Size);
-  Queue.copy(DataB.data(), PtrB, Size);
-  Queue.copy(DataC.data(), PtrC, Size);
-  Queue.wait_and_throw();
-
-  event Event = run_kernels_usm(Queue, Size, PtrA, PtrB, PtrC);
-  Queue.wait_and_throw();
-
-  Graph.begin_recording(Queue);
-  run_kernels_usm(Queue, Size, PtrA, PtrB, PtrC);
-  Graph.end_recording();
-
-  // Execute several iterations of the graph (first iteration has already
-  // run before graph recording)
-  for (unsigned n = 1; n < Iterations; n++) {
-    auto GraphExec = Graph.finalize();
-    Event = Queue.submit([&](handler &CGH) {
-      CGH.depends_on(Event);
-      CGH.ext_oneapi_graph(GraphExec);
-    });
-  }
-  Queue.wait_and_throw();
-
-  Queue.copy(PtrA, DataA.data(), Size);
-  Queue.copy(PtrB, DataB.data(), Size);
-  Queue.copy(PtrC, DataC.data(), Size);
-  Queue.wait_and_throw();
-
-  free(PtrA, Queue);
-  free(PtrB, Queue);
-  free(PtrC, Queue);
-
-  assert(ReferenceA == DataA);
-  assert(ReferenceB == DataB);
-  assert(ReferenceC == DataC);
-
-  return 0;
-}
+#include "../Inputs/multiple_exec_graphs.cpp"
